+++ .gitignore
@@ -23,3 +23,5 @@ npm-debug.log
 .vscode
 *.log
 *.stackdump
+scripts/docs.angularjs.org-firebase/functions/content
+/firebase.json
