+++ Gruntfile.js
@@ -14,7 +14,7 @@ var semver = require('semver');
 var exec = require('shelljs').exec;
 var pkg = require(__dirname + '/package.json');
 
-var docsScriptFolder = 'scripts/docs.angularjs.org-firebase';
+var docsScriptFolder = util.docsScriptFolder;
 
 // Node.js version checks
 if (!semver.satisfies(process.version, pkg.engines.node)) {
@@ -45,7 +45,7 @@ if (!match) {
 }
 
 // Ensure Node.js dependencies have been installed
-if (!process.env.TRAVIS && !process.env.JENKINS_HOME) {
+if (!process.env.CI) {
   var yarnOutput = exec('yarn install');
   if (yarnOutput.code !== 0) {
     throw new Error('Yarn install failed: ' + yarnOutput.stderr);
@@ -109,16 +109,14 @@ module.exports = function(grunt) {
       },
       testserver: {
         options: {
-          // We use end2end task (which does not start the webserver)
-          // and start the webserver as a separate process (in travis_build.sh)
-          // to avoid https://github.com/joyent/libuv/issues/826
+          // We start the webserver as a separate process from the E2E tests
           port: 8000,
           hostname: '0.0.0.0',
           middleware: function(connect, options) {
             var base = Array.isArray(options.base) ? options.base[options.base.length - 1] : options.base;
             return [
               function(req, resp, next) {
-                // cache get requests to speed up tests on travis
+                // cache GET requests to speed up tests
                 if (req.method === 'GET') {
                   resp.setHeader('Cache-control', 'public, max-age=3600');
                 }
@@ -160,8 +158,7 @@ module.exports = function(grunt) {
 
     protractor: {
       normal: 'protractor-conf.js',
-      travis: 'protractor-travis-conf.js',
-      jenkins: 'protractor-jenkins-conf.js'
+      circleci: 'protractor-circleci-conf.js'
     },
 
 
@@ -498,14 +495,9 @@ module.exports = function(grunt) {
     'connect:testserver',
     'protractor:normal'
   ]);
-  grunt.registerTask('test:travis-protractor', 'Run the end to end tests with Protractor for Travis CI builds', [
+  grunt.registerTask('test:circleci-protractor', 'Run the end to end tests with Protractor for CircleCI builds', [
     'connect:testserver',
-    'protractor:travis'
-  ]);
-  grunt.registerTask('test:ci-protractor', 'Run the end to end tests with Protractor for Jenkins CI builds', [
-    'webdriver',
-    'connect:testserver',
-    'protractor:jenkins'
+    'protractor:circleci'
   ]);
   grunt.registerTask('test:e2e', 'Alias for test:protractor', ['test:protractor']);
   grunt.registerTask('test:promises-aplus',[
@@ -538,7 +530,7 @@ module.exports = function(grunt) {
     'package',
     'compress:deployFirebaseCode',
     'copy:deployFirebaseCode',
-    'firebaseDocsJsonForTravis',
+    'firebaseDocsJsonForCI',
     'copy:deployFirebaseDocs'
   ]);
   grunt.registerTask('default', ['package']);
@@ -546,7 +538,7 @@ module.exports = function(grunt) {
 
 
 function reportOrFail(message) {
-  if (process.env.TRAVIS || process.env.JENKINS_HOME) {
+  if (process.env.CI) {
     throw new Error(message);
   } else {
     console.log('===============================================================================');
