--- lib/browserstack/teardown_tunnel.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -e -o pipefail
-
-
-echo "Shutting down Browserstack tunnel"
-echo "TODO: implement me"
-exit 1
