+++ lib/grunt/plugins.js
@@ -63,8 +63,8 @@ module.exports = function(grunt) {
     util.collectErrors();
   });
 
-  grunt.registerTask('firebaseDocsJsonForTravis', function() {
-    util.firebaseDocsJsonForTravis();
+  grunt.registerTask('firebaseDocsJsonForCI', function() {
+    util.firebaseDocsJsonForCI();
   });
 
 };
