+++ lib/grunt/utils.js
@@ -7,6 +7,7 @@ var spawn = require('npm-run').spawn;
 
 var CSP_CSS_HEADER = '/* Include this file in your html if you are using the CSP mode. */\n\n';
 
+const docsScriptFolder = 'scripts/docs.angularjs.org-firebase';
 
 module.exports = {
 
@@ -32,8 +33,8 @@ module.exports = {
 
 
   updateWebdriver: function(done) {
-    if (process.env.TRAVIS) {
-      // Skip the webdriver-manager update on Travis, since the browsers will
+    if (process.env.CI) {
+      // Skip the webdriver-manager update on CI, since the browsers will
       // be provided remotely.
       done();
       return;
@@ -302,11 +303,11 @@ module.exports = {
     };
   },
 
-  // Our Firebase projects are in subfolders, but Travis expects them in the root,
-  // so we need to modify the upload folder path and copy the file into the root
-  firebaseDocsJsonForTravis: function() {
-    var docsScriptFolder = 'scripts/docs.angularjs.org-firebase';
+  docsScriptFolder,
 
+  // Our Firebase projects are in subfolders, but the firebase tool expects them in the root,
+  // so we need to modify the upload folder path and copy the file into the root
+  firebaseDocsJsonForCI: function() {
     var fileName = docsScriptFolder + '/firebase.json';
     var json = grunt.file.readJSON(fileName);
