--- lib/saucelabs/start_tunnel.sh
@@ -1,47 +1,0 @@
-#!/bin/bash
-
-set -e
-
-# Setup and start Sauce Connect for your TravisCI build
-# This script requires your .travis.yml to include the following two private env variables:
-# SAUCE_USERNAME
-# SAUCE_ACCESS_KEY
-# Follow the steps at https://saucelabs.com/opensource/travis to set that up.
-#
-# Curl and run this script as part of your .travis.yml before_script section:
-# before_script:
-#   - curl https://gist.github.com/santiycr/5139565/raw/sauce_connect_setup.sh | bash
-SC_VERSION="4.5.4"
-CONNECT_URL="https://saucelabs.com/downloads/sc-$SC_VERSION-linux.tar.gz"
-CONNECT_DIR="/tmp/sauce-connect-$RANDOM"
-CONNECT_DOWNLOAD="sc-$SC_VERSION-linux.tar.gz"
-
-# We don't want to create a log file because sauceconnect always logs in verbose mode. This seems
-# to be overwhelming Travis and causing flakes when we are cat-ing the log in "print_logs.sh"
-CONNECT_LOG="/dev/null"
-
-# Get Connect and start it
-mkdir -p $CONNECT_DIR
-cd $CONNECT_DIR
-curl $CONNECT_URL -o $CONNECT_DOWNLOAD 2> /dev/null 1> /dev/null
-mkdir sauce-connect
-tar --extract --file=$CONNECT_DOWNLOAD --strip-components=1 --directory=sauce-connect > /dev/null
-rm $CONNECT_DOWNLOAD
-
-SAUCE_ACCESS_KEY=`echo $SAUCE_ACCESS_KEY | rev`
-
-
-ARGS=""
-
-# Set tunnel-id only on Travis, to make local testing easier.
-if [ ! -z "$TRAVIS_JOB_NUMBER" ]; then
-  ARGS="$ARGS --tunnel-identifier $TRAVIS_JOB_NUMBER"
-fi
-if [ ! -z "$BROWSER_PROVIDER_READY_FILE" ]; then
-  ARGS="$ARGS --readyfile $BROWSER_PROVIDER_READY_FILE"
-fi
-
-
-echo "Starting Sauce Connect in the background"
-sauce-connect/bin/sc -u $SAUCE_USERNAME -k $SAUCE_ACCESS_KEY $ARGS \
-  --logfile $CONNECT_LOG &
