--- lib/saucelabs/teardown_tunnel.sh
@@ -1,16 +1,0 @@
-#!/bin/bash
-
-set -e -o pipefail
-
-
-echo "Shutting down Sauce Connect tunnel"
-
-killall sc
-
-while [[ -n `ps -ef | grep "sauce-connect-" | grep -v "grep"` ]]; do
-  printf "."
-  sleep .5
-done
-
-echo ""
-echo "Sauce Connect tunnel has been shut down"
