+++ lib/versions/version-info.js
@@ -65,7 +65,7 @@ var getCodeName = function(tagName) {
 
 
 /**
- * Compute a build segment for the version, from the Jenkins build number and current commit SHA
+ * Compute a build segment for the version, from the CI build number and current commit SHA
  * @return {String} The build segment of the version
  */
 function getBuild() {
@@ -189,7 +189,7 @@ var getSnapshotVersion = function() {
   // We need to clone to ensure that we are not modifying another version
   version = semver(version.raw);
 
-  var jenkinsBuild = process.env.TRAVIS_BUILD_NUMBER || process.env.BUILD_NUMBER;
+  var ciBuild = process.env.CIRCLE_BUILD_NUM || process.env.BUILD_NUMBER;
   if (!version.prerelease || !version.prerelease.length) {
     // last release was a non beta release. Increment the patch level to
     // indicate the next release that we will be doing.
@@ -203,7 +203,7 @@ var getSnapshotVersion = function() {
     // as this is bigger than 1.3.0-beta.2 according to semver
     version.patch++;
   }
-  version.prerelease = jenkinsBuild ? ['build', jenkinsBuild] : ['local'];
+  version.prerelease = ciBuild ? ['build', ciBuild] : ['local'];
   version.build = getBuild();
   version.codeName = 'snapshot';
   version.isSnapshot = true;
