--- scripts/jenkins/init-node.sh
@@ -1,18 +1,0 @@
-#!/bin/bash
-
-# Install nvm for this shell
-source ~/.nvm/nvm.sh
-
-# Use version of node.js found in .nvmrc
-nvm install
-
-# clean out and install yarn
-rm -rf ~/.yarn
-curl -o- -L https://yarnpkg.com/install.sh | bash -s -- --version 1.21.1
-export PATH="$HOME/.yarn/bin:$PATH"
-
-# Ensure that we have the local dependencies installed
-yarn install
-
-echo testing grunt version
-yarn grunt --version
