+++ scripts/release/release.sh
@@ -36,7 +36,6 @@ function init {
 
 function build {
   cd ../..
-  source scripts/jenkins/init-node.sh
   yarn grunt ci-checks package --no-color
 
   cd $SCRIPT_DIR
