--- scripts/travis/tear_down_browser_provider.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-# Has to be run from project root directory.
-
-./lib/${BROWSER_PROVIDER}/teardown_tunnel.sh
